Move public expression classes outside of the catalyst package.
